css/css-contain/content-visibility/content-visibility-size-containment-001.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>content-visibility and layout containment</title>
<link rel="help" href="https://drafts.csswg.org/css-contain/#content-visibility">
<link rel="help" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1765615">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<meta name="assert" content="elements skipping their content change the used value of the contain property to turn on size containment.">
<style>
  /* Change to style containment can cause drastic changes such as rebuilding
     counters or quotes, so always enabled here to prevent any side effect. */
  .content_visibility {
      contain: style;
  }

  #spacer_for_far_to_viewport {
      height: 300vh;
  }

  .content_visibility {
      display: inline-block;
  }
  .box {
      display: inline-block;
      width: 50px;
      height: 5px;
      background: black;
  }

  #visible > .content_visibility {
      content-visibility: visible;
  }
  #hidden > .content_visibility {
      content-visibility: hidden;
  }
  #auto_far > .content_visibility {
      content-visibility: auto;
  }
  #auto_close > .content_visibility {
      content-visibility: auto;
  }
  #visible_to_hidden > .content_visibility {
      content-visibility: visible;
  }
  #hidden_to_visible > .content_visibility {
      content-visibility: hidden;
  }
  #visible_to_auto_far > .content_visibility {
      content-visibility: visible;
  }
  #auto_far_to_visible > .content_visibility {
      content-visibility: auto;
  }
  #hidden_to_auto_close > .content_visibility {
      content-visibility: hidden;
  }
  #auto_close_to_hidden > .content_visibility {
      content-visibility: auto;
  }
</style>
<body>
  <div id="log"></div>

  <div id="visible">
    <div class="box"></div><div class="content_visibility"><div class="box"></div></div><div class="box"></div>
  </div>

  <div id="hidden">
    <div class="box"></div><div class="content_visibility"><div class="box"></div></div><div class="box"></div>
  </div>

  <div id="auto_close">
    <div class="box"></div><div class="content_visibility"><div class="box"></div></div><div class="box"></div>
  </div>

  <div id="visible_to_hidden">
    <div class="box"></div><div class="content_visibility"><div class="box"></div></div><div class="box"></div>
  </div>

  <div id="hidden_to_visible">
    <div class="box"></div><div class="content_visibility"><div class="box"></div></div><div class="box"></div>
  </div>

  <div id="hidden_to_auto">
    <div class="box"></div><div class="content_visibility"><div class="box"></div></div><div class="box"></div>
  </div>

  <div id="auto_to_hidden">
    <div class="box"></div><div class="content_visibility"><div class="box"></div></div><div class="box"></div>
  </div>

  <div id="hidden_to_auto_close">
    <div class="box"></div><div class="content_visibility"><div class="box"></div></div><div class="box"></div>
  </div>

  <div id="auto_close_to_hidden">
    <div class="box"></div><div class="content_visibility"><div class="box"></div></div><div class="box"></div>
  </div>

  <div id="spacer_for_far_to_viewport"></div>

  <div id="auto_far">
    <div class="box"></div><div class="content_visibility"><div class="box"></div></div><div class="box"></div>
  </div>

  <div id="visible_to_auto_far">
    <div class="box"></div><div class="content_visibility"><div class="box"></div></div><div class="box"></div>
  </div>

  <div id="auto_far_to_visible">
    <div class="box"></div><div class="content_visibility"><div class="box"></div></div><div class="box"></div>
  </div>

  <script>
    function sizeContainmentApplied(id) {
        let container = document.getElementById(id);
        let content_visibility = container.getElementsByClassName("content_visibility")[0];
        let beforeBox = content_visibility.previousElementSibling.getBoundingClientRect();
        let afterBox = content_visibility.nextElementSibling.getBoundingClientRect();
        console.log(id)
        console.log(beforeBox)
        console.log(afterBox)
        return afterBox.left - beforeBox.right < 25;
    }

    function setContentVisibility(id, value) {
        let container = document.getElementById(id);
        let content_visibility = container.getElementsByClassName("content_visibility")[0];
        content_visibility.style.contentVisibility = value;
    }

    function tick() {
        return new Promise(resolve => {
            requestAnimationFrame(() => {
                requestAnimationFrame(resolve);
            });
        });
    }

    promise_test(async () => {
        assert_true(!sizeContainmentApplied("visible"));
    }, "size containment does not apply to content-visibility: visible");

    promise_test(async () => {
        assert_true(sizeContainmentApplied("hidden"));
    }, "size containment applies to content-visibility: hidden");

    promise_test(async () => {
        await tick();
        assert_true(sizeContainmentApplied("auto_far"));
    }, "size containment applies to content-visibility: auto (far from viewport)");

    promise_test(async () => {
        await tick();
        assert_true(!sizeContainmentApplied("auto_close"));
    }, "size containment does not apply to content-visibility: auto (close from viewport)");

    promise_test(async () => {
        setContentVisibility("visible_to_hidden", "hidden");
        assert_true(sizeContainmentApplied("visible_to_hidden"));
    }, "size containment updated when switching content-visibility from visible to hidden");

    promise_test(async () => {
        setContentVisibility("hidden_to_visible", "visible");
        assert_true(!sizeContainmentApplied("hidden_to_visible"));
    }, "size containment updated when switching content-visibility from hidden to visible");

    promise_test(async () => {
        setContentVisibility("auto_far_to_visible", "visible");
        assert_true(!sizeContainmentApplied("auto_far_to_visible"));
    }, "size containment updated when switching content-visibility from auto (far from viewport) to visible");

    promise_test(async () => {
        setContentVisibility("visible_to_auto_far", "auto");
        await tick();
        assert_true(sizeContainmentApplied("visible_to_auto_far"));
    }, "size containment updated when switching content-visibility from visible to auto (far from viewport)");

    promise_test(async () => {
        setContentVisibility("auto_close_to_hidden", "hidden");
        assert_true(sizeContainmentApplied("auto_close_to_hidden"));
    }, "size containment updated when switching content-visibility from auto (close from viewport) to hidden");

    promise_test(async () => {
        setContentVisibility("hidden_to_auto_close", "auto");
        await tick();
        assert_true(!sizeContainmentApplied("hidden_to_auto_close"));
    }, "size containment updated when switching content-visibility from hidden to auto (close from viewport)");

  </script>
</body>
